split up zigzag into even more files
