+++ storage-proofs/src/zigzag/porep.rs
@@ -1,0 +1,56 @@
+use crate::error::Result;
+use crate::hasher::Hasher;
+use crate::porep::PoRep;
+use crate::zigzag::{
+    graph::ZigZagBucketGraph,
+    params::{PersistentAux, PublicParams, Tau, TemporaryAux, Tree},
+    proof::ZigZagDrgPoRep,
+};
+
+impl<'a, 'c, H: 'static + Hasher> PoRep<'a, H> for ZigZagDrgPoRep<'a, H> {
+    type Tau = Tau<<H as Hasher>::Domain>;
+    type ProverAux = (PersistentAux<H::Domain>, TemporaryAux<H>);
+
+    fn replicate(
+        pp: &'a PublicParams<H, ZigZagBucketGraph<H>>,
+        replica_id: &<H as Hasher>::Domain,
+        data: &mut [u8],
+        data_tree: Option<Tree<H>>,
+    ) -> Result<(Self::Tau, Self::ProverAux)> {
+        let (tau, p_aux, t_aux) = Self::transform_and_replicate_layers(
+            &pp.graph,
+            &pp.layer_challenges,
+            replica_id,
+            data,
+            data_tree,
+        )?;
+
+        Ok((tau, (p_aux, t_aux)))
+    }
+
+    fn extract_all<'b>(
+        pp: &'b PublicParams<H, ZigZagBucketGraph<H>>,
+        replica_id: &'b <H as Hasher>::Domain,
+        data: &'b [u8],
+    ) -> Result<Vec<u8>> {
+        let mut data = data.to_vec();
+
+        Self::extract_and_invert_transform_layers(
+            &pp.graph,
+            &pp.layer_challenges,
+            replica_id,
+            &mut data,
+        )?;
+
+        Ok(data)
+    }
+
+    fn extract(
+        _pp: &PublicParams<H, ZigZagBucketGraph<H>>,
+        _replica_id: &<H as Hasher>::Domain,
+        _data: &[u8],
+        _node: usize,
+    ) -> Result<Vec<u8>> {
+        unimplemented!();
+    }
+}
